+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-30T08:19:08.752Z",
+    "2026-03-30T08:39:00.224Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {